fix(ci): Include correct requirements and fix Dependabot
I realized that we need to include the honeybee-energy and honeybee-radiance dependency to make it clear that some of the older versions of these packages don't have the schemas of objects here.

I used >= operators since we'll probably only change this maybe once or twice in the life of the package.